consolidate into util.py
